requirements.txt
alembic==1.7.7
anyio==3.6.1
asgiref==3.5.2
click==8.1.3
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.78.0
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.7.1
Mako==1.2.0
MarkupSafe==2.1.1
passlib==1.7.4
psycopg2-binary==2.9.3
pydantic==1.9.1
python-dotenv==0.20.0
python-multipart==0.0.5
PyYAML==6.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.36
starlette==0.19.1
typing_extensions==4.2.0
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
zipp==3.8.0

PyJWT~=2.4.0